Scene transition made with the BallCleaner Script
I couldn't do it with the Scene transition Script, so I called the scene transition function inside the Ball Cleaner, however there are still some bugs and problems, for example, when the ball push ends it still notificating into the log while it's meant to do it only once. The problem is that doing it in ball cleaner restricts me to do a randomization in which scene i'm going to load.